+++ processor/partition_worker_handle_adapter.go
