+++ processor/processor.go
@@ -733,7 +733,7 @@ func (proc *Handle) Start(ctx context.Context) error {
 		proc.logger.Info("Transformer features received")
 
 		h := &workerHandleAdapter{proc}
-		pool := workerpool.New(ctx, func(partition string) workerpool.Worker { return newProcessorWorker(partition, h) }, proc.logger)
+		pool := workerpool.New(ctx, func(partition string) workerpool.Worker { return newPartitionWorker(ctx, partition, h) }, proc.logger)
 		defer pool.Shutdown()
 		for {
 			select {
